Add tests for pseudonymisation  (#214)
* Updated hash and added to db

* Query PIXL db within pixl_dcmd to check dataset exported

* Updated hash_value, started tests

* Remove unused requirements

* Split out deid and datetime helpers

* Use env variables to connect to PIXL db

* Update inputs for tag tests to run

* Negative test for image already exported

* Negative test for image already exported

* Use == instead of is for PIXL query

* Install core in CI

* Check editable install uses dir

* Rollback python version for core

---------
