+++ .github/workflows/main.yml
@@ -112,6 +112,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r pixl_dcmd/src/requirements.txt
+          pip install -e pixl_core
 
       - name: Run tests
         working-directory: pixl_dcmd
