+++ pixl_core/pyproject.toml
@@ -6,7 +6,7 @@ authors = [
 ]
 description = ""
 readme = "README.md"
-requires-python = ">=3.10"
+requires-python = ">=3.9"
 classifiers = [
     "Programming Language :: Python :: 3"
 ]
