+++ pixl_dcmd/bin/run-tests.sh
@@ -18,4 +18,4 @@ BIN_DIR=$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)
 PACKAGE_DIR="${BIN_DIR%/*}"
 cd "$PACKAGE_DIR"
 
-ENV=test pytest src/pixl_dcmd/tests
+pytest src/pixl_dcmd/tests
