+++ pixl_dcmd/src/pixl_dcmd/main.py
@@ -13,19 +13,17 @@
 #  limitations under the License.
 from __future__ import annotations
 
-import hashlib
 import logging
-import re
 from io import BytesIO
 from os import PathLike
-from random import randint
 from typing import Any, BinaryIO, Union
 
-import arrow
 import requests
 from decouple import config
 from pydicom import Dataset, dcmwrite
 from pixl_dcmd._database import insert_new_uid_into_db_entity, query_db
+from pixl_dcmd._deid_helpers import get_bounded_age, get_encrypted_uid
+from pixl_dcmd._datetime import combine_date_time, format_date_time
 
 DicomDataSetType = Union[Union[str, bytes, PathLike[Any]], BinaryIO]
 
@@ -75,108 +73,6 @@ def remove_overlays(dataset: Dataset) -> Dataset:
     return dataset
 
 
-def get_encrypted_uid(uid: str, salt: bytes) -> str:
-    """
-    Hashes the suffix of a DICOM UID with the given salt.
-
-    This function retains the prefix, while sha512-hashing the subcomponents
-    of the suffix. The number of digits per subcomponent is retained in the
-    encrypted UID. This also ensures that no UID is greater than 64 chars.
-    No leading zeros are permitted in a subcomponent unless the subcomponent
-    has a length of 1.
-
-    Original UID:	    1.2.124.113532.10.122.1.203.20051130.122937.2950157
-    Encrypted UID:	1.2.124.113532.74.696.4.703.80155569.949794.5833842
-
-    Encrypting the UIDs this way ensures that no time information remains but
-    that a input UID will always result in the same output UID, for a given salt.
-
-    Note. that while no application should ever rely on the structure of a UID,
-    there is a possibility that the were the anonyimised data to be push to the
-    originating scanner (or scanner type), the data may not be recognised.
-    """
-    uid_elements = uid.split(".")
-
-    prefix = ".".join(uid_elements[:4])
-    suffix = ".".join(uid_elements[4:])
-    logging.debug(f"\t\tPrefix: {prefix}")
-    logging.debug(f"\t\tSuffix: {suffix}")
-
-    # Get subcomponents of suffix as array.
-    suffix_elements = uid_elements[4:]
-    enc_element = [""] * len(suffix_elements)
-
-    # For each subcomponent of the suffix:
-    for idx, item in enumerate(suffix_elements):
-        h = hashlib.sha512()
-        h.update(item.encode("utf-8"))  # Add subcomponent.
-        h.update(salt)  # Apply salt.
-
-        # If subcomponent has a length of one, allow a leading zero, otherwise
-        # strip leading zeros.
-        # Regex removes any non-numeric chars.
-        if len(item) == 1:
-            enc_element[idx] = re.sub("[^0-9]", "", h.hexdigest())[: len(item)]
-        else:
-            enc_element[idx] = re.sub("[^0-9]", "", h.hexdigest()).lstrip("0")[
-                : len(item)
-            ]
-
-    # Return original prefix and encrypted suffix.
-    return prefix + "." + ".".join(enc_element[:])
-
-
-def get_bounded_age(age: str) -> str:
-    """Bounds patient age between 18 and 89"""
-    if age[3] != "Y":
-        return "018Y"
-
-    age_as_int = int(age[0:3])
-    if age_as_int < 18:
-        return "018Y"
-
-    if age_as_int > 89:
-        return "089Y"
-
-    return age
-
-
-def combine_date_time(a_date: str, a_time: str) -> Any:
-    """Turn date string and time string into arrow object."""
-    date_time_str = f"{a_date} {a_time}"
-
-    # TODO: Should Timezone be hardcoded?
-    # https://github.com/UCLH-Foundry/PIXL/issues/151
-    tz = "Europe/London"
-
-    try:
-        new_date_time = arrow.get(date_time_str, tzinfo=tz)
-    except arrow.parser.ParserError:
-        logging.exception(
-            f"Failed to parse the datetime string '{date_time_str}'"
-            f"falling back to a random time in 1970"
-        )
-        new_date_time = arrow.get("1970-01-01T00:00:00+00:00")
-        new_date_time = new_date_time.shift(seconds=randint(10**2, 10**7))
-
-    return new_date_time
-
-
-def format_date_time(a_date_time: str) -> Any:
-    """Turn date-time string into arrow object."""
-    if "." not in a_date_time:
-        a_date_time += ".000000"
-
-    if a_date_time[8] != " ":
-        a_date_time = a_date_time[0:8] + " " + a_date_time[8:]
-
-    if arrow.get(a_date_time, "YYYYMMDD HHmmss.SSSSSS"):
-        a_date = "{s}".format(s=arrow.get(a_date_time).format("YYYYMMDD"))
-        a_time = "{s}".format(s=arrow.get(a_date_time).format("HHmmss.SSSSSS"))
-
-    return combine_date_time(a_date, a_time)
-
-
 def enforce_whitelist(dataset: dict, tags: dict) -> dict:
     """Delete any tags not in the tagging scheme."""
     # For every element:
@@ -215,7 +111,7 @@ def apply_tag_scheme(dataset: dict, tags: dict) -> dict:
     mrn = dataset[0x0010, 0x0020].value  # Patient ID
     accession_number = dataset[0x0008, 0x0050].value  # Accession Number
 
-    salt_plaintext = mrn + accession_number
+    salt_plaintext = config("SALT_VALUE")
 
     HASHER_API_AZ_NAME = config("HASHER_API_AZ_NAME")
     HASHER_API_PORT = config("HASHER_API_PORT")
