+++ pixl_dcmd/src/requirements.txt
@@ -1,9 +1,6 @@
 arrow==1.2.3
 black==22.8.0
 pydicom==2.3.0
-flake8==5.0.4
-flake8-return==1.1.3
-isort==5.10.1
 mypy==0.991
 pytest==7.1.3
 logger==1.4
